feat: Add admin managed users routes and validation for user management
This update introduces new API routes for managing users by organization admins and non-org admins, enhancing user oversight capabilities. Key changes include:

- Added routes for listing, deleting, and managing user geofences.
- Implemented validation to ensure admins can manage the target users.
- Enhanced documentation in CLAUDE.md to reflect the new user management features.

These additions improve the administrative functionality for user management, providing better control and tracking of users within organizations.

+++ crates/api/src/routes/compliance.rs
@@ -180,9 +180,9 @@ async fn generate_compliance_report(
         period_start,
         period_end,
         organization: OrganizationReportSummary {
-            total_devices: dashboard_metrics.devices.total as i64,
-            total_users: dashboard_metrics.users.total as i64,
-            total_groups: dashboard_metrics.groups.total as i64,
+            total_devices: dashboard_metrics.devices.total,
+            total_users: dashboard_metrics.users.total,
+            total_groups: dashboard_metrics.groups.total,
         },
         data_subject_requests: DataSubjectRequestReportSummary {
             total_requests: dsr_counts.total,
